requirements.txt
pybullet>=3.2.5
fastapi>=0.104.1
uvicorn>=0.24.0
websockets>=12.0
pydantic>=2.4.2
python-multipart>=0.0.5
jinja2>=3.0.1
aiofiles>=0.7.0
openai>=1.3.0
python-dotenv>=1.0.0
pytest>=7.4.3
pytest-asyncio>=0.21.1